+++ src/ape/cli/choices.py
@@ -69,8 +69,11 @@ class Alias(click.Choice):
         # NOTE: we purposely skip the constructor of `Choice`
         self.case_sensitive = False
         self._key_filter = key
+
+    @cached_property
+    def choices(self) -> Sequence:  # type: ignore[override]
         module = import_module("ape.types.basic")
-        self.choices = module._LazySequence(self._choices_iterator)
+        return module._LazySequence(self._choices_iterator)
 
     @property
     def _choices_iterator(self) -> Iterator[str]:
@@ -354,18 +357,22 @@ class NetworkChoice(click.Choice):
         base_type: Optional[type] = None,
         callback: Optional[Callable] = None,
     ):
-        provider_module = import_module("ape.api.providers")
-        base_type = provider_module.ProviderAPI if base_type is None else base_type
-        if not issubclass(base_type, (provider_module.ProviderAPI, str)):
-            raise TypeError(f"Unhandled type '{base_type}' for NetworkChoice.")
-
-        self.base_type = base_type
+        self._base_type = base_type
         self.callback = callback
         self.case_sensitive = case_sensitive
         self.ecosystem = ecosystem
         self.network = network
         self.provider = provider
         # NOTE: Purposely avoid super().init for performance reasons.
+
+    def base_type(self) -> type["ProviderAPI"]:
+        if self._base_type is not None:
+            return self._base_type
+
+        from ape.api.providers import ProviderAPI
+
+        self._base_type = ProviderAPI
+        return ProviderAPI
 
     @cached_property
     def choices(self) -> Sequence[Any]:  # type: ignore[override]
@@ -462,3 +469,17 @@ def output_format_choice(options: Optional[list[OutputFormat]] = None) -> Choice
 
     # Uses `str` form of enum for CLI choices.
     return click.Choice([o.value for o in options], case_sensitive=False)
+
+
+class LazyChoice(Choice):
+    """
+    A simple lazy-choice where choices are evaluated lazily.
+    """
+    def __init__(self, get_choices: Callable[[], Sequence[str]], case_sensitive: bool = False):
+        self._get_choices = get_choices
+        self.case_sensitive = case_sensitive
+        # Note: Purposely avoid super init.
+
+    @cached_property
+    def choices(self) -> Sequence[str]:  # type: ignore[override]
+        return self._get_choices()
