+++ src/ape/plugins/compiler.py
@@ -1,6 +1,10 @@
-from ape.api.compiler import CompilerAPI
+from typing import TYPE_CHECKING
 
 from .pluggy_patch import PluginType, hookspec
+
+if TYPE_CHECKING:
+    # perf: Load only for type-checking; makes registering plugins faster.
+    from ape.api.compiler import CompilerAPI
 
 
 class CompilerPlugin(PluginType):
@@ -11,7 +15,9 @@ class CompilerPlugin(PluginType):
     """
 
     @hookspec
-    def register_compiler(self) -> tuple[tuple[str], type[CompilerAPI]]:  # type: ignore[empty-body]
+    def register_compiler(  # type: ignore[empty-body]
+        self,
+    ) -> tuple[tuple[str], type["CompilerAPI"]]:
         """
         A hook for returning the set of file extensions the plugin handles
         and the compiler class that can be used to compile them.
