+++ src/ape/plugins/config.py
@@ -1,6 +1,10 @@
-from ape.api.config import PluginConfig
+from typing import TYPE_CHECKING
 
 from .pluggy_patch import PluginType, hookspec
+
+if TYPE_CHECKING:
+    # perf: Load only for type-checking; makes registering plugins faster.
+    from ape.api.config import PluginConfig
 
 
 class Config(PluginType):
@@ -12,7 +16,7 @@ class Config(PluginType):
     """
 
     @hookspec
-    def config_class(self) -> type[PluginConfig]:  # type: ignore[empty-body]
+    def config_class(self) -> type["PluginConfig"]:  # type: ignore[empty-body]
         """
         A hook that returns a :class:`~ape.api.config.PluginConfig` parser class that can be
         used to deconstruct the user config options for this plugins.
