+++ src/ape_cache/config.py
@@ -1,0 +1,5 @@
+from ape.api.config import PluginConfig
+
+
+class CacheConfig(PluginConfig):
+    size: int = 1024**3  # 1gb
