+++ src/ape_plugins/__init__.py
@@ -1,7 +1,8 @@
-from ape import plugins
-from ape.api.config import ConfigDict
+from ape.plugins import register, Config
 
 
-@plugins.register(plugins.Config)
+@register(Config)
 def config_class():
+    from ape.api.config import ConfigDict
+
     return ConfigDict
